+++ web-console/src/lib/types/connectors.ts
@@ -2,6 +2,7 @@ export enum ConnectorType {
   KAFKA_IN = 'KafkaIn',
   KAFKA_OUT = 'KafkaOut',
   DEBEZIUM_IN = 'DebeziumIn',
+  SNOWFLAKE_OUT = 'SnowflakeOut',
   URL = 'HTTP_GET',
   UNKNOWN = 'Unknown'
 }